requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
bcrypt
pydantic
python-multipart
pyjwt